Design: Edit Board Cell
