+++ TimeLine/Scene/Board/Main/BoardTableViewCell.swift
@@ -15,40 +15,8 @@ final class BoardTableViewCell: UITableViewCell {
     let contentLabel = PlainLabel(size: 13, color: Constants.Color.subText)
     
     let createrLabel = PlainLabel(size: 13, color: Constants.Color.subText)
-    private let reactView = UIView()
-    private lazy var reactStackView = {
-        let view = UIStackView()
-        view.axis = .horizontal
-        view.distribution = .equalSpacing
-        view.alignment = .leading
-        view.spacing = 5
-        view.addArrangedSubview(commentView)
-        view.addArrangedSubview(likeView)
-        return view
-    }()
     
-    private let commentView = UIView()
-    private let likeView = UIView()
-    
-    private let commentImg = {
-        let view = UIImageView()
-        view.image = Constants.Image.comment
-        view.tintColor = Constants.Color.basicText
-        view.contentMode = .scaleAspectFit
-        return view
-    }()
-    
-    private let likeImg = {
-        let view = UIImageView()
-        view.image = Constants.Image.heart
-        view.tintColor = Constants.Color.basicText
-        view.contentMode = .scaleAspectFit
-        return view
-    }()
-    
-    private let commentLabel = PlainLabel(size: 13, color: Constants.Color.subText)
-    private let likeLabel = PlainLabel(size: 13, color: Constants.Color.subText)
-    
+    private let reactView = BoardReactView()
     
     let thumbnailImage = {
         let view = UIImageView()
@@ -92,22 +60,13 @@ final class BoardTableViewCell: UITableViewCell {
         createrLabel.text = nil
         thumbnailImage.isHidden = false
         thumbnailImage.image = nil
-        commentView.isHidden = false
-        likeView.isHidden = false
+        reactView.commentView.isHidden = false
+        reactView.likeView.isHidden = false
     }
     
     private func configure() {
         contentView.addSubview(textStackView)
         contentView.addSubview(thumbnailImage)
-        
-        
-        
-        reactView.addSubview(reactStackView)
-        commentView.addSubview(commentImg)
-        commentView.addSubview(commentLabel)
-        likeView.addSubview(likeImg)
-        likeView.addSubview(likeLabel)
-        
     }
     
     private func setConstraints() {
@@ -131,55 +90,21 @@ final class BoardTableViewCell: UITableViewCell {
         reactView.snp.makeConstraints { make in
             make.height.equalTo(15)
         }
-        reactStackView.snp.makeConstraints { make in
-            make.verticalEdges.leading.equalTo(reactView)
-        }
-        
-        commentView.snp.makeConstraints { make in
-            make.height.equalTo(15)
-            
-        }
-        likeView.snp.makeConstraints { make in
-            make.height.equalTo(15)
-        }
         
         titleLabel.snp.makeConstraints { make in
             make.top.equalTo(textStackView)
         }
-        
-        
-        
-        commentImg.snp.makeConstraints { make in
-            make.leading.verticalEdges.equalTo(commentView)
-            make.size.equalTo(15)
-        }
-        commentLabel.snp.makeConstraints { make in
-            make.verticalEdges.trailing.equalTo(commentView)
-            make.leading.equalTo(commentImg.snp.trailing).offset(4)
-        }
-        
-        likeImg.snp.makeConstraints { make in
-            make.leading.verticalEdges.equalTo(likeView)
-            make.size.equalTo(15)
-        }
-        likeLabel.snp.makeConstraints { make in
-            make.verticalEdges.trailing.equalTo(likeView)
-            make.leading.equalTo(likeImg.snp.trailing).offset(4)
-        }
-    
-        
     }
     
     func setReactLabel(comment: Int, heart: Int) {
         if comment == 0 {
-            commentView.isHidden = true
+            reactView.commentView.isHidden = true
         }
         if heart == 0 {
-            likeView.isHidden = true
+            reactView.likeView.isHidden = true
         }
+        reactView.setLabel(comment, heart)
         
-        commentLabel.text = "\(comment)"
-        likeLabel.text = "\(heart)"
     }
